gitstream/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
yt-dlp>=2024.5.27
google-generativeai==0.7.0
supabase==2.4.6
python-dotenv==1.0.1
mangum==0.17.0
